Data.Semiring

Description

Categorical semirings (my term, but maybe the correct one) are an alternative to Arrows, but play nicely with groupoids. | See the source or the latex source for more background.

Endofunctors for construction

class Category k => Ctor k constr | constr -> k where

An endofunctor for combining two morphisms|

Methods

selfmap :: k a b -> k c d -> k (constr a c) (constr b d)

first/right like functions

promote :: Ctor k op => k a b -> k (op a c) (op b c)

construct a new morphism with identity|

swap_promote :: Ctor k op => k a b -> k (op c a) (op c b)

construct a new morphism with identity with the arguments reversed|

Laws (Axioms) for building algebraic structures

class Ctor k op => Absorbs k op id | op -> k, op -> id where

The absorbtion law => x+0 <-> x |

Methods

absorb :: k (op id a) a

unabsorb :: k a (op id a)

class Ctor k op => Assocative k op | op -> k where

The assocative law => (x + y) + z <-> x + (y + z)|

Methods

assoc :: k (op (op a b) c) (op a (op b c))

unassoc :: k (op a (op b c)) (op (op a b) c)

class Ctor k op => Commutative k op | op -> k where

The commutative law => x + y <-> y + x|

Methods

commute :: k (op a b) (op b a)

class Ctor k op => Annihilates k op zero | op zero -> k where

The annihilation law => 0 * x <-> 0|

Methods

annihilates :: k (op zero a) zero

class (Ctor k add, Ctor k multi) => Distributes k add multi | add multi -> k where

The distribution law => (a + b) * c <-> (a * c) + (b * c)|

Methods

distribute :: k (multi (add a b) c) (add (multi a c) (multi b c))

undistribute :: k (add (multi a c) (multi b c)) (multi (add a b) c)

Categorical Algebraic Structures

class (Assocative k dot, Absorbs k dot id) => Monoidial k dot id | dot id -> k

Monoidial Category class|

class (Monoidial k dot id, Commutative k dot) => CommutativeMonoidial k dot id | dot id -> k

Commutative Monoidial Category class|

class (CommutativeMonoidial k add zero, CommutativeMonoidial k multi one, Annihilates k multi zero, Distributes k add multi) => Semiring k add zero multi one | add zero multi one -> k

Semiring Category class|

Arrow like functions for semiring categories

first :: Semiring a add zero multi one => a b c -> a (multi b d) (multi c d)

Apply the multi monoid operator to the morphism and identity|

second :: Semiring a add zero multi one => a b c -> a (multi d b) (multi d c)

Apply the multi monoid operator to identity and the morphism|

left :: Semiring a add zero multi one => a b c -> a (add b d) (add c d)

Apply the add monoid operator to the morphism and identity|

right :: Semiring a add zero multi one => a b c -> a (add d b) (add d c)

Apply the add monoid operator to identity and the morphism|

A groupoid class that is a category. Maybe this is a bad idea?

class Category g => Groupoid g where

Methods

inv :: g a b -> g b a

data Iso k a b

Constructors

Iso

Fields

embed :: k a b

project :: k b a

Alegraic laws as isomorphism for groupoid instances

biject_sum_absorb :: Biject (BSum BZero a) a

biject_sum_assoc :: Biject (BSum (BSum a b) c) (BSum a (BSum b c))

biject_product_absorb :: Biject (BProduct BOne a) a

biject_product_assoc :: Biject (BProduct (BProduct a b) c) (BProduct a (BProduct b c))

biject_distributes :: Biject (BProduct (BSum a b) c) (BSum (BProduct a c) (BProduct b c))

kbiject_sum_absorb :: Monad m => KBiject m (KBSum KBZero a) a

kbiject_sum_assoc :: Monad m => KBiject m (KBSum (KBSum a b) c) (KBSum a (KBSum b c))

kbiject_product_absorb :: Monad m => KBiject m (KBProduct KBOne a) a

kbiject_product_assoc :: Monad m => KBiject m (KBProduct (KBProduct a b) c) (KBProduct a (KBProduct b c))

kbiject_distributes :: Monad m => KBiject m (KBProduct (KBSum a b) c) (KBSum (KBProduct a c) (KBProduct b c))